requirements.txt
snakemake
pandas
numpy
biopython
matplotlib
ipykernel
requests
py3Dmol